Working on DLL
The input system is going to be a bit tricky, so I want to set up the DLL code first, then get back to that. The DLL code won't be tested until later, but I want to at least set it all the functions now. So far, the only functions I'm missing are the file writing and reading functions.

For the reading and writing, I plan on listing the size of each piece of data first (separated by a new line), adding some character to represent a split, then putting in the data itself. I figured I'd do it this way so that I don't have to limit the data size.