Export Works - Import Is Still a Problem
So, calling to delete all data records causes the program to crash. Maybe it has something to do with having the memory allocated on the Unity side? 
Anyway, the export works, so now I'm doing the import. The import seems to work properly on the C++ side, but restoring the information is an issue.