Import Seems to Work?
The import of the data appears to work well enough. I'm still worried about it screwing up, but that shouldn't be a problem for the demonstration at least.

I intend on doing the bonus, but I'm not sure if I'll have time.  I can do the procedural generation, which would just reuse a past assignment I've done, but I don't know how well that would work.