Base LevelGenerator from In-Class 6 Added
The base level generator from in-class 6 has been added. I'm likely going to change it to just generate positions, which are then used to move around existing cubes. I also have the ground tagged since it will need to be made larger, and the player is tagged so that they don't get caught up with the randomization. The player will still need to be placed somehow though, which will probably be a random spot on the grid at the highest elevation.